move datetime into a new module
This change creates datetime into a new mod, which is only included and
used when crono is disabled.
